requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
requests>=2.25.0
beautifulsoup4>=4.9.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
arxiv>=1.4.0
feedparser>=6.0.0
PyGithub>=1.55
python-dateutil>=2.8.0
pydantic>=1.8.0
tqdm>=4.62.0
jupyter>=1.0.0
plotly>=5.0.0
dash>=2.0.0
transformers>=4.10.0
torch>=1.9.0 